debug suite au merge

+++ libft/ft_strinsert.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strinsert.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rficht <robin.ficht@free.fr>               +#+  +:+       +#+        */
+/*   By: mdjemaa <mdjemaa@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/09 16:45:28 by rficht            #+#    #+#             */
-/*   Updated: 2023/05/09 17:04:02 by rficht           ###   ########.fr       */
+/*   Updated: 2023/05/11 12:09:15 by mdjemaa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	ft_strinsert(char *str1, char *str2, int pos)
 		return (1);
 	i = -1;
 	j = -1;
-	if (pos > ft_strlen(str1))
+	if ((size_t) pos > ft_strlen(str1))
 		return (1);
 	result = ft_calloc((ft_strlen(str1) + ft_strlen(str2)), sizeof(char));
 	while (++i < pos)
